Made a potential contact list.
Certain trigger updates were causing potentialEnterObject to be called
inside a container query, so when a raycast check was used, the engine
spat the dummy.

Now all visibility checking happens inside processTick. Objects discovered
using potentialEnterObject are pushed onto a holding list which is combed
for valid contacts in processTick.